% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{calculate_interchrom}
\alias{calculate_interchrom}
\title{Calculate inter-chromosomal contacts given walks using one-shot formula}
\usage{
calculate_interchrom(
  intra.dts,
  walk.dts,
  tiled.target,
  walk.cn,
  if.comps = F,
  mc.cores = 1,
  if.fast = T
)
}
\arguments{
\item{intra.dts}{list of data.tables of intra-chromosomal contacts for each walk, generated by simulate_map}

\item{walk.dts}{list of data.tables corresponding to each walk, with at least one column orig.id}

\item{tiled.target}{data.table of target region tiles and their coordinates, given by prep_for_sim()}

\item{walk.cn}{array giving the CN of each walk in walk.dts}

\item{if.comps}{boolean, whether or not to consider compartments (defaults to FALSE)}
}
\value{
single data.table corresponding to the total contacts in the target region including both intra- and inter-chromosomal contacts
}
\description{
Calculate inter-chromosomal contacts given walks using one-shot formula
}
