% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastKar-package.R
\docType{package}
\name{fastKar-package}
\alias{fastKar}
\alias{fastKar-package}
\title{fastKar: Simulate rearranged karyotype Hi-C, fast!}
\description{
This package implements a probabilistic model of Hi-C data for highly rearranged genomes, using the gGnome framework developed by the Imielinski Lab.
}
\author{
\strong{Maintainer}: Iraj Eshghi \email{ieshghi@nygenome.org}

}
\keyword{internal}
