% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{make_template_dat}
\alias{make_template_dat}
\title{Given a set of bins, make the gMatrix data.table corresponding to all pairwise contacts for those bins}
\usage{
make_template_dat(target.bins, if.comps = FALSE)
}
\arguments{
\item{target.bins}{either a GRanges, or a data.table. If DT, must have a column "width", giving the width of all the bins}
}
\value{
data.table with columns i,j,id,widthprod corresponding to all pairwise contacts, widthprod is the product of widths of tiles i and j
}
\description{
Given a set of bins, make the gMatrix data.table corresponding to all pairwise contacts for those bins
}
