% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{prep_for_sim}
\alias{prep_for_sim}
\title{Tile target region and prepare data necessary for Hi-C simulation.}
\usage{
prep_for_sim(walks, target_region = NULL, pix.size = 1e+05, if.comps = F)
}
\arguments{
\item{walks}{either a gGnome::gWalk or a list containing objects "graph" (a gGnome gGraph), "snode.id" (a list of arrays of node.ids inside "graph" in their walk order), and "circular" (an array of booleans indicating the circularity of the walks)}

\item{target_region}{a gRange corresponding to the region of interest for the simulation. If left blank, assume the footprint of the walks.}

\item{pix.size}{resolution of the simulation tiles, default is 100kb. If set to 0, each tile corresponds to half a node in the gGraph.}

\item{if.comps}{boolean setting whether or not to simulate compartments (defaults to FALSE)}
}
\value{
a list with two elements: "tiled.target", a data.table containing the tiles, their coordinates, and compartment identity, and "widthdt", a data.table containing the widths of all nodes from the input gGraph
}
\description{
Tile target region and prepare data necessary for Hi-C simulation.
}
