% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{simulate_map}
\alias{simulate_map}
\title{Simulate Hi-C data using spline trained on Rao 2014 data}
\usage{
simulate_map(walk.dt, is.circular = FALSE, model = 0)
}
\arguments{
\item{walk.dt}{data.table prepared by simulate_walks(), each row is a tile to simulate and columns are start,end,midpoints of the tile in local walk coordinates.}

\item{is.circular}{boolean indicating whether the walk is circular or not (defaults to FALSE)}

\item{model}{sets whether to simulate Hi-C (model = 0) or long-read data (model = length of reads in bp)}
}
\value{
data.table of long-format gMatrix with columns i,j,id,value
}
\description{
Simulate Hi-C data using spline trained on Rao 2014 data
}
