% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{simulate_walks}
\alias{simulate_walks}
\title{Downstream of prep_for_sim, calculate Hi-C data in reference coordinates given sets of walks on a graph and a tiled target region}
\usage{
simulate_walks(
  walks,
  tiled.target,
  widthdt,
  if.comps = F,
  mc.cores = 1,
  if.sum = T,
  depth = 1,
  model = 0,
  if.interchr = T,
  gm.out = T
)
}
\arguments{
\item{walks}{either a gGnome::gWalk or a list containing objects "graph" (a gGnome gGraph), "snode.id" (a list of arrays of node.ids inside "graph" in their walk order), and "circular" (an array of booleans indicating the circularity of the walks)}

\item{tiled.target}{a data.table produced by prep_for_sim(), containing target tile coordinates, tile.ids, and corresponding node.ids, as well as CN and width}

\item{widthdt}{a data.table produced by prep_for_sim(), containing width data for all the nodes in the gGraph simulated here.}

\item{if.comps}{boolean setting whether or not to simulate compartments (defaults to FALSE)}

\item{mc.cores}{number of cores to use in parallel simulation (defaults to 1)}

\item{if.sum}{boolean deciding whether to sum the simulated Hi-C signal from all walks or whether to return a list of Hi-C maps (defaults to TRUE)}

\item{depth}{depth of the simulated Hi-C data, in X coverage (defaults to 1)}

\item{model}{sets whether to simulate Hi-C (model = 0) or long-read data (model = length of reads in bp)}

\item{if.interchr}{boolean, sets whether to simulate interchromosomal contacts (default is TRUE). If if.sum = FALSE, then this is set to FALSE.}

\item{gm.out}{boolean, chooses whether to return a gMatrix or a data.table (defaults to TRUE, returning a gMatrtix)}
}
\value{
either a gMatrix (gm.out = if.sum = T), a list of gMatrices (gm.out = T, if.sum = F), a data.table (gm.out = F, if.sum = T), or a list of data.tables (gm.out = if.sum = F)
}
\description{
Downstream of prep_for_sim, calculate Hi-C data in reference coordinates given sets of walks on a graph and a tiled target region
}
