% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{sum_matrices}
\alias{sum_matrices}
\title{Sums the contacts coming from many gMatrix data.tables as fast as possible using the 'id' column.}
\usage{
sum_matrices(matrices)
}
\arguments{
\item{matrices}{list of data.tables of long-form gMatrices with columns i,j,id,value}
}
\value{
single data.table with same columns, with value column corresponding to the sum of all inputs for a given id
NOTE: could be sped up with cpp
}
\description{
Sums the contacts coming from many gMatrix data.tables as fast as possible using the 'id' column.
}
