% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{symmetrize}
\alias{symmetrize}
\title{Symmetrize a matrix}
\usage{
symmetrize(input.mat)
}
\arguments{
\item{input.mat}{some input matrix, must be square}
}
\value{
matrix of the same shape as input.mat, with same diagonal, and non-diagonal being sum of matrix and its transpose
NOTE: could be sped up with cpp
}
\description{
Symmetrize a matrix
}
