% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward_simulate}
\alias{forward_simulate}
\title{Simulate the Hi-C data of a set of gWalks in reference coordinates using the fastKar framework}
\usage{
forward_simulate(
  walks,
  target_region = NULL,
  pix.size = 1e+05,
  if.comps = FALSE,
  mc.cores = 1,
  if.sum = TRUE,
  depth = 1,
  model = 0,
  if.interchr = T,
  gm.out = T
)
}
\arguments{
\item{walks}{either a gGnome::gWalk or a list containing objects "graph" (a gGnome gGraph), "snode.id" (a list of arrays of node.ids inside "graph" in their walk order), and "circular" (an array of booleans indicating the circularity of the walks)}

\item{target_region}{a gRange corresponding to the region of interest for the simulation. If left blank, assume the footprint of the walks.}

\item{pix.size}{resolution of the simulation tiles, default is 100kb. If set to 0, each tile corresponds to half a node in the gGraph.}

\item{if.comps}{boolean setting whether or not to simulate compartments (defaults to FALSE)}

\item{mc.cores}{number of cores to use in parallel simulation (defaults to 1)}

\item{if.sum}{boolean deciding whether to sum the simulated Hi-C signal from all walks or whether to return a list of Hi-C maps (defaults to TRUE)}

\item{depth}{depth of the simulated Hi-C data, in X coverage (defaults to 1)}

\item{model}{sets whether to simulate Hi-C (model = 0) or long-read data (model = length of reads in bp)}

\item{if.interchr}{boolean, sets whether to simulate interchromosomal contacts (default is TRUE). If if.sum = FALSE, then this is set to FALSE.}

\item{gm.out}{boolean, chooses whether to return a gMatrix or a data.table (defaults to TRUE, returning a gMatrtix)}
}
\value{
either a gMatrix (gm.out = if.sum = T), a list of gMatrices (gm.out = T, if.sum = F), a data.table (gm.out = F, if.sum = T), or a list of data.tables (gm.out = if.sum = F)
}
\description{
Simulate the Hi-C data of a set of gWalks in reference coordinates using the fastKar framework
}
